+++ src/containers/AuthPage/constants.js
@@ -1,4 +1,26 @@
-export const VK_URL = 'https://oauth.vk.com/authorize';
-export const EACH_URL = 'http://each.itsociety.su:5000/oauth2/authorize?response_type=code&client_id=Gu2SCEBUwQV3TSlNIu8uMzvKRMYuGP5ePh044jGErO6O9RR0&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fauth&scope=email&state=TW9uIE9jdCAwOCAyMDE4IDE4OjU2OjAwIEdNVCswMzAwICjQnNC%2B0YHQutCy0LAsINGB0YLQsNC90LTQsNGA0YLQvdC%2B0LUg0LLRgNC10LzRjyk%3D';
-export const GOOGLE_URL = 'https://accounts.google.com/o/oauth2/auth';
-export const GOOGLE_REDIRECT_URL = 'https://auth.expo.io/auth'; // 'http://each.itsociety.su:10000/auth';
+import { AuthSession } from 'expo';
+
+const googleClientId = '190923403189-srp0gleu6imvtph8gcauf03uhb66q65h.apps.googleusercontent.com';
+const eachClientId = 'Gu2SCEBUwQV3TSlNIu8uMzvKRMYuGP5ePh044jGErO6O9RR0';
+const vkontakteClientId = '6682398';
+const redirectUrl = AuthSession.getRedirectUrl();
+
+export const googleAuthUrl = `https://accounts.google.com/o/oauth2/auth?` +
+  `&client_id=${encodeURIComponent(googleClientId)}` +
+  `&redirect_uri=${encodeURIComponent(redirectUrl)}` +
+  `&response_type=code` +
+  `&scope=${encodeURIComponent(['profile','email'].join(' '))}`;
+
+export const eachAuthUrl = `http://each.itsociety.su:5000/oauth2/authorize?` +
+  `&client_id=${encodeURIComponent(eachClientId)}` +
+  `&redirect_uri=${encodeURIComponent(redirectUrl)}` +
+  `&response_type=code` +
+  `&scope=${encodeURIComponent(['email'].join(' '))}`;
+
+export const vkontakteAuthUrl = `https://oauth.vk.com/authorize?` +
+  `&client_id=${encodeURIComponent(vkontakteClientId)}` +
+  `v=${encodeURIComponent("5.85")}` +
+  `display=mobile` +
+  `&redirect_uri=${encodeURIComponent(redirectUrl)}` +
+  `&response_type=code` +
+  `&scope=${encodeURIComponent(['offline','email'].join(' '))}`;
